chore: refactor and tests
