+++ packages/design-system/src/utils/functions.ts
@@ -1,4 +1,4 @@
-import { FlatList, Linking, Platform } from 'react-native'
+import { Linking, Platform } from 'react-native'
 
 export const openLink = (link: string, newTab = false) => {
   if (Platform.OS === 'web') {
@@ -14,24 +14,3 @@ export const openLink = (link: string, newTab = false) => {
 
 export const isLast = (index: number, array: unknown[]) =>
   index === array.length - 1
-
-const SCROLL_RETRY_TIMEOUT = 500
-export const makeScrollToIndexFailedHandler = (
-  listRef: React.RefObject<Pick<FlatList<any>, 'scrollToIndex'>>,
-) => {
-  let attemptsLeft = 3
-  let timeoutId: NodeJS.Timeout | undefined
-  return (info: { index: number; highestMeasuredFrameIndex: number }) => {
-    if (timeoutId) {
-      clearTimeout(timeoutId)
-    }
-    if (attemptsLeft === 0) {
-      return
-    }
-    timeoutId = setTimeout(() => {
-      listRef.current?.scrollToIndex({ index: info.index })
-      timeoutId = undefined
-      attemptsLeft -= 1
-    }, SCROLL_RETRY_TIMEOUT)
-  }
-}
